requirements.txt
asgiref==3.5.2
Django==3.2.16
django-admin-sortable2==1.0.4
django-tinymce==3.5.0
environs==9.5.0
Pillow==9.3.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
